% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{add}
\usage{
add(files = NULL, ...)
}
\arguments{
\item{files}{Character vector of file(s) to add.}

\item{...}{Currently ignored. For extendability.}
}
\value{
Adds files
}
\description{
Add a file
}
\examples{
\dontrun{
file.create("new_file.txt")
add("new_file.txt")
}
}
