% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_all_to_git_stage.R
\name{add_all_to_git_stage}
\alias{add_all_to_git_stage}
\title{Add ALL files in the RStudio Git pane to the 'stage'
for comitting}
\usage{
add_all_to_git_stage()
}
\description{
Add ALL files in the RStudio Git pane to the 'stage'
for comitting
}
