% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-checkout.R
\name{checkout}
\alias{checkout}
\alias{git_checkout}
\title{Git Checkout}
\usage{
git_checkout(branch = NULL, file = NULL)
}
\arguments{
\item{branch}{\code{character(1)}. The name of a branch, typically a
feature branch.}

\item{file}{\code{character(1)}. The path to a file.}
}
\value{
\code{NULL} ... invisibly.
}
\description{
Checks out as a branch if doesn't exist. Branch
oriented workflow for switching between branches.
If \code{file} is passed, checks out the file. A common
shortcut to undoing local changes to a file(s).
Can be a vector of multiple files.
}
\examples{
\dontrun{
  git_checkout("feature-br")

  git_checkout(file = "DESCRIPTION")
}
}
