% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.R
\name{commit}
\alias{commit}
\title{Commit}
\usage{
commit(message, files = NULL)
}
\arguments{
\item{message}{Commit message}

\item{files}{File(s) to commit}

\item{...}{Currently ignored. For extendability.}
}
\description{
Commit a file
}
