\name{contextMenu}
\alias{contextMenu}
\title{Add a context menu}
\usage{
  contextMenu(h, widget, event, action = NULL, ...)
}
\arguments{
  \item{h}{handler list}

  \item{widget}{(unused)}

  \item{event}{gtkEvent}

  \item{action}{(unused)}
}
\description{
  Function that adds Context Menu like add3rdMousePopupMenu
  but with dynamic menu.
}

