% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_git_user.R
\docType{data}
\name{email}
\alias{email}
\title{A function to easily set the git user.name and email in R}
\format{An object of class \code{character} of length 1.}
\usage{
email
}
\description{
A function to easily set the git user.name and email in R
}
\keyword{datasets}
