\name{expandRows}
\alias{expandRows}
\title{Expand Rows}
\usage{
  expandRows(obj, iter, expandedRows, root = NULL)
}
\arguments{
  \item{obj}{gTreeRGtk object}

  \item{iter}{iter thingy}

  \item{expandedRows}{vector of paths to expand}

  \item{root}{used in recursive calling of the function}
}
\description{
  Restore the saved expansion state.
}

