% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{Gadd}
\alias{Gadd}
\title{\code{git add --all}, aka, quick-add}
\usage{
Gadd(...)
}
\arguments{
\item{...}{Objects to add; uses '--all' if nothing provided}
}
\value{
Any results from running \code{git add --all}
}
\description{
\code{git add --all}, aka, quick-add
}
\examples{
# Gadd()
}

