% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{Gamend}
\alias{Gamend}
\title{\code{git commit --amend -m 'x'} with message x}
\usage{
Gamend(msg)
}
\arguments{
\item{msg}{A message to include with a commit}
}
\value{
Any results from running \code{git commit --amend -m}
}
\description{
\code{git commit --amend -m 'x'} with message x
}
\examples{
# Gamend("A silly commit message.")
}

