% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{Gcommit}
\alias{Gcommit}
\title{\code{git commit -m 'x'} with message x}
\usage{
Gcommit(msg)
}
\arguments{
\item{msg}{A message to include with a commit}
}
\value{
Any results from running commit
}
\description{
\code{git commit -m 'x'} with message x
}
\examples{
# Gcommit("A silly commit message.")
}

