% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{Ginit}
\alias{Ginit}
\title{\code{git init} a new repo}
\usage{
Ginit(force = FALSE)
}
\arguments{
\item{force}{Whether to force init [FALSE]}
}
\value{
Any results from running \code{git init}
}
\description{
Checks with user that the init is being done in the right directory, to avoid
those cases where the user isn't where they think they are. Also checks that
there isn's a .git directory first so that a repo isn't re-inited #' (unless
\code{force = TRUE}).
}
\examples{
# Ginit()
}

