% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{Git}
\alias{Git}
\title{An empty \code{git} call}
\usage{
Git(cmd = "")
}
\arguments{
\item{cmd}{A message to include with a commit}
}
\value{
Any results from running the command
}
\description{
To be used if a git command doesn't already exist in this package.
}
\examples{
# git("add .")
}

