% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{Gpush}
\alias{Gpush}
\title{\code{git push} to a remote repo}
\usage{
Gpush()

Gpush()
}
\value{
Any results from running the command

Any results from running \code{git pull}
}
\description{
\code{git push} to a remote repo

\code{git pull} from a remote repo
}
\examples{
# Gpush()
# Gpull()
}

