% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{Gqa}
\alias{Gqa}
\title{\code{git add --all && commit -m 'x'} with message x}
\usage{
Gqa(msg)
}
\arguments{
\item{msg}{A message to include with a commit}
}
\value{
Any results from running commit
}
\description{
\code{git add --all && commit -m 'x'} with message x
}
\examples{
# Gqa("A silly commit message.")
}

