% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{Gstatus}
\alias{Gstatus}
\title{\code{git status}}
\usage{
Gstatus()
}
\arguments{
\item{msg}{A message to include with a commit}
}
\value{
Any results from running commit
}
\description{
\code{git status}
}
\examples{
# Gstatus()
}

