% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitr-params.R
\name{params}
\alias{params}
\title{Common Parameters for gitr}
\arguments{
\item{n}{\code{integer(1)}. How far back to go from current \code{HEAD}.
Same as the command line \verb{git log -n} parameter.
For \verb{git stash} commands, zero-index into the stash list.}

\item{file}{\code{character(1)}. The path to a file.}

\item{branch}{\code{character(1)}. The name of a branch, typically a
feature branch.}

\item{sha}{\code{character(n)}. The commit secure hash algorithm (SHA-1).
If \code{NULL}, typically points to the most recent commit on
the current branch.}
}
\description{
Common Parameters for gitr
}
