% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull}
\alias{pull}
\title{pull}
\usage{
pull(...)
}
\arguments{
\item{...}{Currently ignored. For extendability.}
}
\value{
Pulls changes not yet on local device
}
\description{
Pull
}
\examples{
\dontrun{
pull()
}
}
