% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{push}
\alias{push}
\title{push}
\usage{
push(...)
}
\arguments{
\item{...}{Currently ignored. For extendability.}
}
\value{
Pushes local commits
}
\description{
Push local commits
}
\examples{
\dontrun{
file.create("new_file.txt")
commit("new_file.txt")
push()
}
}
