% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_git_user.R
\name{set_git_user}
\alias{set_git_user}
\title{A function to easily set the git user.name and email in R}
\usage{
set_git_user(email, username)
}
\description{
A function to easily set the git user.name and email in R
}
