% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitr-sitrep.R
\name{sitrep}
\alias{sitrep}
\alias{gitr_sitrep}
\title{Git Situation Report}
\usage{
gitr_sitrep()
}
\value{
\code{NULL} ... invisibly.
}
\description{
Get a situation report of the current git repository.
}
