% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-branch.R
\name{branch}
\alias{branch}
\alias{git_default_br}
\alias{git_current_br}
\title{Git Branch Utilities}
\usage{
git_default_br()

git_current_br()
}
\value{
Character. The name of the respective branch if found, otherwise \code{NULL}.
}
\description{
Git Branch Utilities
}
\section{Functions}{
\itemize{
\item \code{git_default_br()}: gets the default "main" branch, typically either \code{master}, \code{main}, or \code{trunk}.

\item \code{git_current_br()}: gets the \emph{current} branch.

}}
