% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{Ginit}
\alias{Ginit}
\title{\code{git init} a new repo}
\usage{
Ginit(force = FALSE)
}
\arguments{
\item{force}{Whether to force init [FALSE]}
}
\value{
Any results from running \code{git init}
}
\description{
Checks that there isn's a .git directory first so that a repo isn't re-inited
(unless you force).
}
\examples{
# Ginit()
}

