% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitr-package.R
\docType{package}
\name{gitr-package}
\alias{gitr}
\alias{gitr-package}
\title{gitr: A Lightweight API for Git}
\description{
A light-weight, dependency-free, API to access system-level Git commands from within R. Contains wrappers and defaults for common data science workflows as well as Z-shell plugin aliases. Generalized API syntax is also available. A system installation of Git is required.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stufield.github.io/gitr/}
  \item Report bugs at \url{https://github.com/stufield/gitr/issues}
}

}
\author{
\strong{Maintainer}: Stu Field \email{stu.g.field@gmail.com} [copyright holder]

}
\keyword{internal}
