% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitr-package.R
\docType{package}
\name{gitr-package}
\alias{gitr}
\alias{gitr-package}
\title{gitr: A Lightweight API for 'Git'}
\description{
A light-weight, dependency-free, application programming interface (API) to access system-level 'Git' \url{https://git-scm.com/downloads} commands from within 'R'. Contains wrappers and defaults for common data science workflows as well as 'Zsh' \url{https://github.com/ohmyzsh/ohmyzsh} plugin aliases. A generalized API syntax is also available.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stufield.github.io/gitr/}
  \item Report bugs at \url{https://github.com/stufield/gitr/issues}
}

}
\author{
\strong{Maintainer}: Stu Field \email{stu.g.field@gmail.com} (\href{https://orcid.org/0000-0002-1024-5859}{ORCID}) [copyright holder]

}
\keyword{internal}
