% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-lint-commit.R
\name{lint}
\alias{lint}
\alias{lint_commit_msg}
\title{Common Lints for Commit Messages}
\usage{
lint_commit_msg(x)
}
\arguments{
\item{x}{A single commit message from \code{\link[=get_commit_msgs]{get_commit_msgs()}}.}
}
\value{
\code{integer(1)}. Invisibly returns the
number of detected lints in the message.
}
\description{
Lint a commit message for typical commit style and
best practices for git.
}
\examples{
\dontrun{
  lapply(get_commit_msgs(7L), lint_commit_msg)
}
}
