% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Common Parameters for gitr}
\arguments{
\item{n}{Numeric. How far back to go from current \code{HEAD}.
Same as the command line \verb{git log -n} parameter.
For \verb{git stash} commands, zero-index into the stash list.}

\item{file}{Character. The path name to a file.}

\item{branch}{Character. The name of a branch, typically a
feature branch.}

\item{sha}{Character. The commit SHA-1 hash to pull messages from.
If \code{NULL}, the most recent commit on the current branch.}
}
\description{
Common Parameters for gitr
}
