% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-pull-request.R
\name{pr}
\alias{pr}
\alias{get_pr_msgs}
\alias{get_pr_sha}
\title{Git PR Utilities}
\usage{
get_pr_msgs(branch = NULL)

get_pr_sha(branch = NULL)
}
\arguments{
\item{branch}{Character. The path name to a file.}
}
\description{
Git PR Utilities
}
\section{Functions}{
\itemize{
\item \code{get_pr_msgs()}: Gets the commit messages for the \emph{current} branch relative to
the \verb{origin/\{main,master\}} branch in the remote. Typically these "new" commits
that would be merged as part of a PR to \verb{origin/\{main,master\}}.

\item \code{get_pr_sha()}: Gets the commit SHA1 \emph{current} branch relative to
the \code{default} branch in the remote, usually either \code{origin/main} or
\code{origin/master}. See \code{\link[=git_default_br]{git_default_br()}}.

}}
