% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim-sha.R
\name{sha}
\alias{sha}
\alias{trim_sha}
\title{SHA1 Utilities}
\usage{
trim_sha(sha)
}
\arguments{
\item{sha}{Character. The commit SHA-1 hash to pull messages from.
If \code{NULL}, the most recent commit on the current branch.}
}
\value{
Character. The trimmed \code{sha}.
}
\description{
SHA1 Utilities
}
\section{Functions}{
\itemize{
\item \code{trim_sha()}: trims the \code{SHA-1} hash from the default full length to the
human-readable short version.

}}
