% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-tag.R
\name{tag}
\alias{tag}
\alias{git_recent_tag}
\alias{git_tag_info}
\title{Git Tag Utilities}
\usage{
git_recent_tag()

git_tag_info()
}
\description{
Git Tag Utilities
}
\section{Functions}{
\itemize{
\item \code{git_recent_tag()}: Get the \emph{most} recent \code{git} tag.

\item \code{git_tag_info()}: Gets a data frame summary of the current git repository tags.

}}
