% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitr-tag.R
\name{tag}
\alias{tag}
\alias{gitr_recent_tag}
\alias{gitr_tag_info}
\title{Git Tag Utilities}
\usage{
gitr_recent_tag()

gitr_tag_info()
}
\value{
\code{\link[=gitr_recent_tag]{gitr_recent_tag()}}: \code{character(1)}. The most recent tag.

\code{\link[=gitr_tag_info]{gitr_tag_info()}}: A data frame summarizing the repository tags.
}
\description{
Git Tag Utilities
}
\section{Functions}{
\itemize{
\item \code{gitr_recent_tag()}: gets the \emph{most} recent \code{git} tag.

\item \code{gitr_tag_info()}: gets a data frame summary of
the current git repository tags.

}}
\examples{
\dontrun{
  gitr_recent_tag()

  gitr_tag_info()
}
}
