% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsh.R
\name{zsh}
\alias{zsh}
\alias{glog}
\alias{gco}
\alias{gpr}
\alias{gss}
\alias{gba}
\alias{gau}
\alias{gpop}
\alias{gta}
\alias{gfa}
\alias{gac}
\alias{gwip}
\alias{gclean}
\alias{gdf}
\alias{gnuke}
\alias{gcf}
\title{Z-shell Aliases}
\usage{
glog(n = 10)

gco(branch = NULL)

gpr()

gss()

gba()

gau()

gpop()

gta()

gfa()

gac()

gwip()

gclean(dry.run = TRUE)

gdf(file = NULL)

gnuke()

gcf(global = FALSE)
}
\arguments{
\item{n}{Numeric. How far back to go from current HEAD. Same as the
command line \verb{git log -n} parameter.}

\item{branch}{Character. The name of a branch, typically a
feature branch.}

\item{dry.run}{Logical. Clean as dry-run?}

\item{file}{A full file path within the repository to diff.}

\item{global}{Logical. Query global repository. Alternatively local config.}
}
\description{
Provides ZSH-like aliases to common Z-shell plugin aliases.
}
\section{Functions}{
\itemize{
\item \code{glog}: Get the \code{git} log in a pretty format for the \code{n} most recent commits.

\item \code{gco}: \verb{git checkout}.

\item \code{gpr}: \verb{git pull --rebase}.

\item \code{gss}: \verb{git status -s}.

\item \code{gba}: \verb{git branch -a}.

\item \code{gau}: \verb{git add -u}.

\item \code{gpop}: \verb{git stash pop}.

\item \code{gta}: \verb{git tag -n}.

\item \code{gfa}: \verb{git fetch --all --prune}.

\item \code{gac}: \verb{git commit --amend --no-edit}.

\item \code{gwip}: \verb{git commit -m 'wip'}.

\item \code{gclean}: \verb{git clean -f -d}.

\item \code{gdf}: \verb{git diff <file>}.

\item \code{gnuke}: \verb{git reset --hard}.

\item \code{gcf}: \verb{git config --local} or \verb{git config --global}.
}}

\examples{
\dontrun{
glog()
}
}
